Prac12_MQTT/MQTT_Mongodb/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
eventlet==0.36.1
fastapi==0.110.1
Flask==3.0.3
Flask-SocketIO==5.3.6
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
logging==0.4.9.6
MarkupSafe==2.1.5
paho-mqtt==2.0.0
pika==1.3.2
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
Werkzeug==3.0.2
wsproto==1.2.0
